pyproject.toml
[tool.poetry]
name = "twacapic"
version = "0.1.4.5"
description = "A Twitter Academic API Client"
authors = ["Felix Victor Münch <f.muench@leibniz-hbi.de>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Leibniz-HBI/twacapic"
classifiers = [
		"Development Status :: 2 - Pre-Alpha",
		"Environment :: Console",
		"Intended Audience :: Science/Research",
]
exclude = [
		"twacapic/twacapic_credentials.txt",
		"*/.pytest_cache/*",
		"twitter_keys.yaml",
		"results"
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
TwitterAPI = "^2.6.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
coverage = "^5.5"

[tool.poetry.scripts]
twacapic = 'twacapic.main:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
